time_unit_step_size
- inorder to stay consistent with the year and month type time windows,
time_unit_step_size is used if time_window_type in
set_analysis_parameters is 'range'. This is analogous to the number of
days in year. e.g. a time_unit_step_size=24 and ws=240 mean to have a
window of 240 steps and in the next iteration move the window 24 steps.
- not fully tested. Comparision with year type analysis shows identical
results. But testing of other outputs and flags is not done yet.
- some new additional private variables i.e. _ws_inc and _tuss were
added to cope with the range time window analysis.